chaos: effect of changing x_0
This is the logistic map. The graph of the function fmap tends to show chaotic behaviour when a > 3.45. Here it's set to 3.74, so chaotic behaviour is expected. 
Note that as the initial condition, x_0, is changed (x_0 = 0.2, in this case), no two graphs will be similar/the same. The graphs with very similar initial conditions may start off on the same trajectory, but after many iterations, k, the difference between the two functions may be very different. This is shown (not in this code, however) on a logaithmic relationship between the distance between the divergence of the two graphs. After many k (or iterations), the natural log graph will be close to 0, which suggests exponential divergence.